Refactor of Managers folder content into a "Pipeline"
