+++ src/BlazorGPT.Web/Pages/HelloWorldPage.razor
@@ -1,5 +1,5 @@
 ﻿@page "/helloworld"
-@using BlazorGPT.Managers
+@using BlazorGPT.Pipeline
 @layout EmptyLayout
 <h3>Hello World</h3>
