+++ src/BlazorGPT.Web/Pages/Pinball.razor
@@ -1,6 +1,6 @@
 ﻿@page "/pinball"
 @page "/pinball/{table}"
-@using BlazorGPT.Managers
+@using BlazorGPT.Pipeline
 @layout EmptyLayout
 <h3>PinballGPT</h3>
