+++ src/BlazorGPT.Web/Pages/SharePage.razor
@@ -1,6 +1,6 @@
 ﻿@page "/share/{conversationId:guid?}"
-@using BlazorGPT.Managers
 @using Microsoft.EntityFrameworkCore
+@using BlazorGPT.Pipeline
 @layout EmptyLayout
 <h3>@Conversation?.Summary</h3>
