+++ src/BlazorGPT.Web/appsettings.Development.json
@@ -8,7 +8,7 @@
   },
   "PipelineOptions": {
     "KrokiHost": "http://localhost:8001",
-    "EnabledInterceptors": [ "Json Hive State", "Save file", "Structurizr Hive DSL", "State has changed" ]
-
+    "EnabledInterceptors": [ "Json Hive State", "Save file", "Structurizr Hive DSL", "State has changed" ],
+    "StateFileSaveInterceptorPath": "C:\\source\\BlazorGPT\\src\\BlazorGPT\\wwwroot\\state"
   }
 }
