+++ src/BlazorGPT/BlazorGPT.csproj
@@ -46,7 +46,6 @@
 
   <ItemGroup>
     <Folder Include="Migrations\" />
-    <Folder Include="Pipeline\" />
     <Folder Include="wwwroot\state\messages\" />
   </ItemGroup>
