+++ src/BlazorGPT/Data/SampleDataSeeder.cs
@@ -1,5 +1,5 @@
 ﻿using BlazorGPT.Data.Model;
-using BlazorGPT.Managers;
+using BlazorGPT.Pipeline;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 using OpenAI.GPT3.ObjectModels.RequestModels;
