+++ src/BlazorGPT/Pages/GPTModelSelector.razor
@@ -1,5 +1,5 @@
 @using System.Collections
-@using BlazorGPT.Managers
+@using BlazorGPT.Pipeline
 @using Microsoft.Extensions.Options
 @inject IOptions<PipelineOptions> PipelineOptions
