+++ src/BlazorGPT/Pipeline/PipelineOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace BlazorGPT.Managers;
+﻿namespace BlazorGPT.Pipeline;
 
 public class PipelineOptions
 {
@@ -6,4 +6,5 @@ public class PipelineOptions
     public string[] Models { get; set; }
     public string[]? EnabledInterceptors { get; set; }
     public string KrokiHost { get; set; }
+    public string StateFileSaveInterceptorPath { get; set; }
 }
