+++ src/BlazorGPT/Shared/StructurizrViewer.razor
@@ -1,7 +1,7 @@
 ﻿@using System.Net.Http.Json
 @using System.Text
 @using BlazorGPT.Data
-@using BlazorGPT.Managers
+@using BlazorGPT.Pipeline
 @using Microsoft.Extensions.Options
 @using Newtonsoft.Json
